fix: fixed issue where a re-opened zaak could not be closed by adding a check when updating a zaak result to see if there was an existing result or not

+++ src/main/java/net/atos/client/zgw/shared/ZGWApiService.java
@@ -152,8 +152,10 @@ public class ZGWApiService {
      * @return Created {@link Resultaat}.
      */
     public Resultaat updateResultaatForZaak(final Zaak zaak, final UUID resultaatTypeUuid, final String reden) {
-        final Resultaat resultaat = zrcClientService.readResultaat(zaak.getResultaat());
-        zrcClientService.deleteResultaat(resultaat.getUuid());
+        if (zaak.getResultaat() != null) {
+            final Resultaat resultaat = zrcClientService.readResultaat(zaak.getResultaat());
+            zrcClientService.deleteResultaat(resultaat.getUuid());
+        }
         return createResultaatForZaak(zaak, resultaatTypeUuid, reden);
     }
